+++ linopy/expressions.py
@@ -11,6 +11,7 @@ import logging
 from dataclasses import dataclass, field
 from itertools import product, zip_longest
 from typing import Any, Mapping, Optional, Union
+from warnings import warn
 
 import numpy as np
 import pandas as pd
@@ -187,7 +188,7 @@ class LinearExpressionGroupby:
             coords = Coordinates.from_pandas_multiindex(idx, group_dim)
             ds = self.data.assign_coords(coords)
             ds = ds.unstack(group_dim, fill_value=LinearExpression._fill_value)
-            ds = LinearExpression._sum(ds, dims=GROUPED_TERM_DIM)
+            ds = LinearExpression._sum(ds, dim=GROUPED_TERM_DIM)
 
             if int_map is not None:
                 index = ds.indexes["group"].map({v: k for k, v in int_map.items()})
@@ -276,7 +277,7 @@ class LinearExpression:
 
     Summation over dimensions
 
-    >>> type(expr.sum(dims="dim_0"))
+    >>> type(expr.sum(dim="dim_0"))
     <class 'linopy.expressions.LinearExpression'>
     """
 
@@ -671,27 +672,27 @@ class LinearExpression:
         return sol.rename("solution")
 
     @classmethod
-    def _sum(cls, expr: Union["LinearExpression", Dataset], dims=None) -> Dataset:
+    def _sum(cls, expr: Union["LinearExpression", Dataset], dim=None) -> Dataset:
         data = _expr_unwrap(expr)
 
-        if dims is None:
+        if dim is None:
             vars = DataArray(data.vars.data.ravel(), dims=TERM_DIM)
             coeffs = DataArray(data.coeffs.data.ravel(), dims=TERM_DIM)
             const = data.const.sum()
             ds = xr.Dataset({"vars": vars, "coeffs": coeffs, "const": const})
         else:
-            dims = [d for d in np.atleast_1d(dims) if d != TERM_DIM]
+            dim = [d for d in np.atleast_1d(dim) if d != TERM_DIM]
             ds = (
                 data[["coeffs", "vars"]]
-                .reset_index(dims, drop=True)
+                .reset_index(dim, drop=True)
                 .rename({TERM_DIM: STACKED_TERM_DIM})
-                .stack({TERM_DIM: [STACKED_TERM_DIM] + dims}, create_index=False)
+                .stack({TERM_DIM: [STACKED_TERM_DIM] + dim}, create_index=False)
             )
-            ds["const"] = data.const.sum(dims)
+            ds["const"] = data.const.sum(dim)
 
         return ds
 
-    def sum(self, dims=None, drop_zeros=False) -> "LinearExpression":
+    def sum(self, dim=None, drop_zeros=False, **kwargs) -> "LinearExpression":
         """
         Sum the expression over all or a subset of dimensions.
 
@@ -699,16 +700,27 @@ class LinearExpression:
 
         Parameters
         ----------
-        dims : str/list, optional
+        dim : str/list, optional
             Dimension(s) to sum over. The default is None which results in all
             dimensions.
+        dims : str/list, optional
+            Deprecated. Use ``dim`` instead.
 
         Returns
         -------
         linopy.LinearExpression
             Summed expression.
         """
-        res = self.__class__(self._sum(self, dims=dims), self.model)
+        if dim is None and "dims" in kwargs:
+            dim = kwargs.pop("dims")
+            warn(
+                "The `dims` argument in `.sum` is deprecated. Use `dim` instead.",
+                DeprecationWarning,
+            )
+        if kwargs:
+            raise ValueError(f"Unknown keyword argument(s): {kwargs}")
+
+        res = self.__class__(self._sum(self, dim=dim), self.model)
 
         if drop_zeros:
             res = res.densify_terms()
@@ -1379,10 +1391,10 @@ class QuadraticExpression(LinearExpression):
         return sol.rename("solution")
 
     @classmethod
-    def _sum(cls, expr: "QuadraticExpression", dims=None) -> Dataset:
+    def _sum(cls, expr: "QuadraticExpression", dim=None) -> Dataset:
         data = _expr_unwrap(expr)
-        dims = dims or list(set(data.dims) - set(HELPER_DIMS))
-        return LinearExpression._sum(expr, dims)
+        dim = dim or list(set(data.dims) - set(HELPER_DIMS))
+        return LinearExpression._sum(expr, dim)
 
     def to_constraint(self, sign, rhs):
         raise NotImplementedError(