+++ doc/release_notes.rst
@@ -14,6 +14,12 @@ Upcoming Version
 
 * Up to now the `rhs` argument in the `add_constraints` function was not supporting an expression as an input type. This is now added.
 
+* Linopy now supports python 3.12.
+
+**Deprecations**
+
+* The argument `dims` in the `.sum` function of variables and expressions was deprecated in favor of the `dim` argument. This aligns the argument name with the xarray convention.
+
 Version 0.3.8
 -------------